requirements.txt
Django==2.2
psycopg2-binary==2.7.7
Pillow>=5.4.1
django-import-export>=1.2.0